fixing things
